fix: case fixes
Related work item: GH-7
