+++ src/components/PasswordInput/PasswordInput.tsx
@@ -1,5 +1,4 @@
 import '../../i18n/i18n';
-import './PasswordInput.scss';
 import { IconButton, InputAdornment, TextField } from '@mui/material';
 import { Visibility, VisibilityOff } from '@mui/icons-material';
 import { Field } from 'formik';
