+++ src/components/RegisterForm/RegisterForm.tsx
@@ -1,33 +1,36 @@
 import '../../i18n/i18n';
-import { Box, Button, Divider, Grid } from '@mui/material';
-import { Err } from './RegisterForm.types';
-import { Formik } from 'formik';
-import { RegisterInput } from '../RegisterInput/RegisterInput';
+import { Box, Button, Divider, Grid, TextField, useMediaQuery, useTheme } from '@mui/material';
+import { Field, Formik } from 'formik';
+import { PasswordInput } from '../PasswordInput/PasswordInput';
+import { RegisterFormErrors } from './RegisterForm.types';
+import styles from './RegisterForm.module.scss';
 import { useTranslation } from 'react-i18next';
 
 export function RegisterForm() {
   const minLength = 1;
   const { t } = useTranslation();
   const validate = (values: { email: string; password: string; lname: string; fname: string; phone: string }) => {
-    const errors: Err = {};
+    const errors: RegisterFormErrors = {};
     if (values.email.length < minLength) {
-      errors.email = t('RegisterForm.email_empty');
+      errors.Email = t('registerForm.emailEmpty');
     }
     if (values.password.length < minLength) {
-      errors.password = t('RegisterForm.password_empty');
+      errors.Password = t('registerForm.passwordEmpty');
     }
     if (values.lname.length < minLength) {
-      errors.lname = t('RegisterForm.last_name_empty');
+      errors.LastName = t('registerForm.lastNameEmpty');
     }
     if (values.fname.length < minLength) {
-      errors.fname = t('RegisterForm.first_name_empty');
+      errors.FirstNme = t('registerForm.firstNameEmpty');
     }
     if (values.phone.length < minLength) {
-      errors.phone = t('RegisterForm.phone_empty');
+      errors.Phone = t('registerForm.phoneEmpty');
     }
     return errors;
   };
 
+  const theme = useTheme();
+  const isMobile: boolean = useMediaQuery(theme.breakpoints.down('tablet'));
   return (
     <Formik
       enableReinitialize
@@ -71,7 +74,89 @@ export function RegisterForm() {
                   mobile={10}
                   tablet={11}
                 >
-                  <RegisterInput />
+                  <Grid
+                    container
+                    alignItems='center'
+                    justifyContent='center'
+                    spacing={{
+                      desktop: 4,
+                      mobile: 2,
+                      tablet: 4,
+                    }}
+                  >
+                    <Grid
+                      item
+                      tablet={6}
+                      mobile={12}
+                    >
+                      <Field
+                        as={TextField}
+                        className={styles.input}
+                        label={t('registerForm.firstName')}
+                        type='text'
+                        name='fname'
+                        required
+                      />
+                    </Grid>
+                    <Grid
+                      item
+                      tablet={6}
+                      mobile={12}
+                    >
+                      <Field
+                        as={TextField}
+                        className={styles.input}
+                        label={t('registerForm.lastName')}
+                        type='text'
+                        name='lname'
+                        required
+                      />
+                    </Grid>
+                    <Grid
+                      item
+                      tablet={6}
+                      mobile={12}
+                    >
+                      <Field
+                        as={TextField}
+                        className={styles.input}
+                        label={t('registerForm.phone')}
+                        inputProps={{ inputMode: 'numeric', pattern: '[0-9]*' }}
+                        type='text'
+                        name='phone'
+                        required
+                      />
+                    </Grid>
+                    {!isMobile && (
+                      <Grid
+                        item
+                        tablet={6}
+                        mobile={0}
+                      ></Grid>
+                    )}
+
+                    <Grid
+                      item
+                      tablet={6}
+                      mobile={12}
+                    >
+                      <Field
+                        as={TextField}
+                        className={styles.input}
+                        label={t('registerForm.email')}
+                        type='email'
+                        name='email'
+                        required
+                      />
+                    </Grid>
+                    <Grid
+                      item
+                      tablet={6}
+                      mobile={12}
+                    >
+                      <PasswordInput />
+                    </Grid>
+                  </Grid>
                 </Grid>
                 <Grid
                   item
@@ -94,7 +179,7 @@ export function RegisterForm() {
                     type='submit'
                     disabled={!isValid}
                   >
-                    {t('RegisterForm.register')}
+                    {t('registerForm.register')}
                   </Button>
                 </Grid>
               </Grid>
