+++ src/components/RegisterForm/RegisterForm.types.ts
@@ -1,7 +1,7 @@
-export type Err = {
-  email?: string;
-  password?: string;
-  lname?: string;
-  fname?: string;
-  phone?: string;
+export type RegisterFormErrors = {
+  Email?: string;
+  Password?: string;
+  LastName?: string;
+  FirstNme?: string;
+  Phone?: string;
 };
