--- src/components/RegisterInput/RegisterInput.scss
@@ -1,3 +1,0 @@
-.input {
-  width: 100%;
-}
