+++ src/locales/en/registerForm.json
@@ -1,14 +1,14 @@
 {
   "password": "Password",
   "phone": "Phone Number",
-  "first_name": "First Name",
-  "last_name": "Last Name",
+  "firstName": "First Name",
+  "lastName": "Last Name",
   "email": "Email",
   "register": "Register",
-  "email_empty": "Email is required",
-  "password_empty": "Password is required",
-  "phone_empty": "Phone is required",
-  "first_name_empty": "First Name is required",
-  "last_name_empty": "Last Name is required",
-  "show_password": "Show Password"
+  "emailEmpty": "Email is required",
+  "passwordEmpty": "Password is required",
+  "phoneEmpty": "Phone is required",
+  "firstNameEmpty": "First Name is required",
+  "lastNameEmpty": "Last Name is required",
+  "showPassword": "Show Password"
 }
