--- src/locales/pl/RegisterForm.json
@@ -1,14 +1,0 @@
-{
-    "password": "Hasło",
-    "phone": "Numer Telefonu",
-    "first_name": "Imie",
-    "last_name": "Nazwisko",
-    "email": "Email",
-    "register": "Zarejestruj Się",
-    "email_empty": "Email jest wymagany",
-    "password_empty": "Hasło jest wymagane",
-    "phone_empty": "Numer telefonu jest wymagany",
-    "first_name_empty": "Imie jest wymagane",
-    "last_name_empty": "Nazwisko jest wymagane",
-    "show_password": "Pokaż hasło"
-}
