+++ src/pages/RegisterPage.tsx
@@ -1,0 +1,5 @@
+import { RegisterForm } from '../components/RegisterForm/RegisterForm';
+
+export function RegisterPage() {
+  return <RegisterForm />;
+}
